Implement draft of a more compact IntMap
The trick is to store keys in inner nodes and use them instead of the prefix
to store the prefix that is shared by all keys at this node and the nodes
below. Getting rid of the prefix and one level of leaf-nodes reduces the
storage overhead from about 6 words per node to only about 2.5 words per node.
The inner nodes are selected such that they have the smallest key of all the
nodes sharing the prefix. This is necessary to still get an efficient
`toAscList` and it has the nice benefit that findMin becomes O(1).

I have not done any benchmarking yet, but I expect this structure as fast as
the existing IntMap's, as we're doing less allocation, but a bit more
operations per node.