Speed up equality; remove expensive invariant
We used to define `==` to test equality in *priority* order. This was
expensive in and of itself, since it takes logarithmic time to produce
each entry. But the situation was worse than that: since priorities are
not unique, we introduced an extra invariant to make sure that keys with
the same priority are extracted in key order. This requires additional
comparisons in various performance-critical functions.

* Remove the extra invariant.
* Remove the extra code to preserve it.
* Change `(==)` to work in key order instead of priority order.
  Keys are unique, so this is very natural.
* Remove a bunch of now-redundant constraints.
* Document the newly introduced nondeterminism.