Tighten OrdPSQ balance property test.
This change tightens the test condition to exactly match the balance
property described in the Hinze and Adams papers. This causes the
Quickcheck test for tree validity to fail in most runs. This is at
least somewhat expected, as it has been established that Adams'
balancing algorithm fails in some edge cases, and Hinze's paper
was written before that discovery.